Inline goal window helper and drop PR-only tests
