+++ GameSentenceMiner/web/goals_api.py
@@ -7,6 +7,7 @@ Provides data for calculating progress on user-defined goals with date ranges.
 
 import datetime
 import json
+from dataclasses import dataclass
 import pytz
 import time
 from flask import request, jsonify
@@ -15,7 +16,6 @@ from GameSentenceMiner.util.config.configuration import logger
 from GameSentenceMiner.util.database.db import GameLinesTable, GoalsTable
 from GameSentenceMiner.util.database.games_table import GamesTable
 from GameSentenceMiner.util.database.stats_rollup_table import StatsRollupTable
-from GameSentenceMiner.util.goals_progress import calculate_goal_progress_window
 from GameSentenceMiner.web.rollup_stats import (
     calculate_live_stats_for_today,
     aggregate_rollup_data,
@@ -26,6 +26,33 @@ from GameSentenceMiner.web.rollup_stats import (
 
 
 # Helper Functions
+
+
+@dataclass(frozen=True)
+class GoalProgressWindow:
+    include_today_live: bool
+    rollup_end_date: datetime.date
+    has_rollup_range: bool
+
+
+def calculate_goal_progress_window(
+    start_date: datetime.date, end_date: datetime.date, today: datetime.date
+) -> GoalProgressWindow:
+    """
+    Calculate the historical/live stats window for a goal.
+
+    Historical rollups are capped at the goal's own end date. Live stats are
+    only included while the goal still overlaps today.
+    """
+    include_today_live = end_date >= today
+    yesterday = today - datetime.timedelta(days=1)
+    rollup_end_date = min(end_date, yesterday) if include_today_live else end_date
+
+    return GoalProgressWindow(
+        include_today_live=include_today_live,
+        rollup_end_date=rollup_end_date,
+        has_rollup_range=start_date <= rollup_end_date,
+    )
 
 
 def combine_stats_with_third_party(rollup_stats, live_stats, start_date_str=None, end_date_str=None):
